refactor(icons): vend "<icons-demo>" element for use with iconsets
