+++ packages/icons/stories/icons.stories.ts
@@ -10,65 +10,12 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 import { storiesOf } from '@storybook/polymer';
-import { html, TemplateResult } from 'lit-html';
+import { html } from 'lit-html';
 
 import '../';
 import '../../icon';
-import { IconsetAddedDetail } from '../../iconset';
-import { LitElement, css, CSSResult } from 'lit-element';
+import '../../iconset/lib/icons-demo';
 import { color } from '@storybook/addon-knobs';
-
-class IconsDemo extends LitElement {
-    private iconset: string[] = [];
-    public constructor() {
-        super();
-        this.iconset = [];
-        this.handleIconSetAdded = this.handleIconSetAdded.bind(this);
-    }
-    public connectedCallback(): void {
-        super.connectedCallback();
-        window.addEventListener('sp-iconset-added', this.handleIconSetAdded);
-    }
-    public handleIconSetAdded(event: CustomEvent<IconsetAddedDetail>): void {
-        const { iconset } = event.detail;
-        this.iconset = iconset.getIconList();
-        this.requestUpdate();
-    }
-    public static get styles(): CSSResult[] {
-        return [
-            css`
-                :host {
-                    display: grid;
-                    grid-template-columns: repeat(auto-fit, minmax(150px, 1fr));
-                    gap: 20px;
-                }
-                .icon {
-                    display: flex;
-                    flex-direction: column;
-                    align-items: center;
-                }
-                sp-icon {
-                    margin-bottom: 10px;
-                }
-            `,
-        ];
-    }
-    protected render(): TemplateResult {
-        return html`
-            <slot></slot>
-            ${this.iconset.map(
-                (icon) => html`
-                    <div class="icon">
-                        <sp-icon size="l" name=${`ui:${icon}`}></sp-icon>
-                        ${icon}
-                    </div>
-                `
-            )}
-        `;
-    }
-}
-
-customElements.define('icons-demo', IconsDemo);
 
 storiesOf('Icons', module)
     .add(
